Add support for --externalserverip
When running servers and registering them with a central server,
Jamulus will auto-detect an external IP of the current machine.
However, this will not be a publicly reachable IP when using NAT and a
central server in the same private network.
This means, that the server is inaccessible via that central server.

Therefore, add a new command line option --externalserverip to override
this auto-detection.

Note: This still requires that proper port forwarding is set up. It also
assumes that ports are forwarded symmetrically, i.e.
	external-ip:22124 -> internal-ip:22124

Fixes #888.

Signed-off-by: Christian Hoffmann <mail@hoffmann-christian.info>