Add support for central servers/slaves behind NAT
Previously, the server list handed out the originating address
of a slave server to any client.
A special case was implemented to hand out internal addresses
instead of public addresses when both the slave server and the
client are assumed to be behind the same NAT.

This commit adds another special case for running central servers
and slave servers behind NAT while serving clients from the Internet.
In such a setup, slave servers would only be known by their internal
address to the central server. As such, clients would be served an
internal IP address which they cannot use.

When slave servers make use of the newly added --externalserverip flag,
they can now override their internal IP address with the proper public
IP address.

This commit adapts the server list logic to make use of an IP which is
provided in this way.

Note: This feature is only available when the code is built by Qt 5.11
or newer.

Fixes #888.

Signed-off-by: Christian Hoffmann <mail@hoffmann-christian.info>